requirements.txt
bcrypt==3.1.4
cffi==1.11.5
Django==2.0.6
djangorestframework==3.8.2
psycopg2==2.7.4
pycparser==2.18
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
